Move sglang_client.py into common and update reference in opd_reward.py. Remove min_containers=1 for serve_sglang.py standalone deployments.
